requirements.txt
duckdb
deltalake
getdaft[deltalake]
numpy
pandas